{"type":"TI", "runs":"0"}

+++ exerciseCh5.java
@@ -82,7 +82,7 @@ public class exerciseCh5 extends ConsoleProgram {
 	
 	private double raiseRealToPower (double x, int k) {
 		double result=1;
-		int fraction = 1;
+		double fraction = 1;
 		if (k<0) {
 			for (int i=0;i>k;i--) {
 				fraction = (int) fraction * x;
